moving Ranges to spectra-data-ranges
