+++ packages/spectra-data-ranges/.eslintrc.yml
@@ -1,1 +1,5 @@
 extends: 'eslint-config-cheminfo'
+env:
+  mocha: true
+rules:
+  valid-jsdoc: error
