--- packages/spectra-data-ranges/.travis.yml
@@ -1,6 +1,0 @@
-language: node_js
-node_js:
-  - node
-  - lts/boron # v6
-  - lts/argon # v4
-
