+++ packages/spectra-data-ranges/package.json
@@ -32,5 +32,9 @@
     "mocha": "^3.2.0",
     "mocha-better-spec-reporter": "^3.1.0",
     "should": "^11.1.2"
+  },
+  "dependencies": {
+    "ml-stat": "^1.3.0",
+    "spectra-nmr-utilities": "^0.0.2"
   }
 }
