+++ packages/spectra-data-ranges/src/index.js
@@ -1,3 +1,4 @@
 'use strict';
 
-module.exports = spectraDataRanges;
+module.exports.GUI = require('./visualizer/annotations');
+module.exports.Ranges = require('./range/Ranges');
