+++ packages/spectra-data/src/index.js
@@ -3,5 +3,5 @@
 exports.SD = require('./SD');
 exports.NMR = require('./NMR');
 exports.NMR2D = require('./NMR2D');
-exports.Ranges = require('./range/Ranges');
-exports.GUI = require('./range/visualizer/index');
+exports.Ranges = require('spectra-data-ranges').Ranges;
+exports.GUI = require('spectra-data-ranges').GUI;
