--- packages/spectra-data/src/range/__tests__/acsTest.json
@@ -1,75 +1,0 @@
-[
-  {
-    "from": 7.259480829364015,
-    "to": 7.32372921947523,
-    "integral": 1.946980644294626,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "m",
-        "kind": "",
-        "remark": "",
-        "delta": 7.282535423195773
-      }
-    ],
-    "signalID": "1",
-  },
-  {
-    "from": 7.147953157908694,
-    "to": 7.234259379817134,
-    "integral": 2.809723174640099,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "m",
-        "delta": 7.203205792670993
-      }
-    ],
-    "signalID": "2"
-  },
-  {
-    "from": 2.562625091175312,
-    "to": 2.6449298824014726,
-    "integral": 1.9683026078604824,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "q",
-        "j": [
-          {
-            "coupling": 7.616016909087193,
-            "multiplicity": "q"
-          }
-        ],
-        "delta": 2.6037774867883923
-      }
-    ],
-    "signalID": "3"
-  },
-  {
-    "from": 1.1584526860710453,
-    "to": 1.2228060847121707,
-    "integral": 2.8260362213770875,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "t",
-        "j": [
-          {
-            "coupling": 7.6310209819084776,
-            "multiplicity": "t"
-          }
-        ],
-        "delta": 1.1906293853916081
-      }
-    ],
-    "signalID": "4",
-    "_highlight": [
-      "4"
-    ]
-  }
-]
