--- packages/spectra-data/src/range/acs/__tests__/ranges.json
@@ -1,174 +1,0 @@
-[
-  {
-    "from": 2.9207911403931033,
-    "to": 3.005342719673898,
-    "integral": 3.05464,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "dd",
-        "kind": "",
-        "remark": "",
-        "delta": 2.9630669300,
-        "assignment": "COCH3",
-        "j": [
-          {
-            "coupling": 2.7202333074737908,
-            "multiplicity": "d"
-          },
-          {
-            "coupling": 9.176268879054533,
-            "multiplicity": "d"
-          }
-        ]
-      }
-    ]
-  },
-  {
-    "from": 1.2768799268371258,
-    "to": 1.3121201215410911,
-    "integral": 1.2945124056168826,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "t",
-        "kind": "",
-        "remark": "",
-        "delta": 1.29450,
-        "j": [
-          {
-            "coupling": 4.531576643179051,
-            "multiplicity": "t"
-          }
-        ]
-      }
-    ]
-  },
-  {
-    "from": 3.2022,
-    "to": 3.5045,
-    "integral": 3.231412397661983,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "d",
-        "kind": "",
-        "delta": 3.3541,
-        "remark": "",
-        "pubAssignment": "CH3OH",
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          }
-        ]
-      }
-    ]
-  },
-  {
-    "from": 4.0935,
-    "to": 4.341,
-    "integral": 3.231412397661983,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "d",
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          }
-        ],
-        "delta": 4.2341
-      },
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "t",
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          },
-          {
-            "coupling": 5.5,
-            "multiplicity": "d"
-          }
-        ],
-        "delta": 4.154
-      }
-    ]
-  },
-  {
-    "from": 4.86,
-    "to": 4.79123,
-    "integral": 3.231412397661983,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "multiplicity": "d",
-        "kind": "",
-        "remark": "",
-        "pubAssignment": "D2O",
-        "delta": 4.80465,
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          }
-        ]
-      }
-    ]
-  },
-  {
-    "from": 5.4234,
-    "to": 5.6423,
-    "integral": 3.231412397661983,
-    "signal": [
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          }
-        ],
-        "delta": 5.462
-      },
-      {
-        "nbAtoms": 0,
-        "diaID": [],
-        "kind": "",
-        "remark": "",
-        "j": [
-          {
-            "coupling": 5.194226872123636,
-            "multiplicity": "d"
-          },
-          {
-            "coupling": 5.5,
-            "multiplicity": "d"
-          }
-        ],
-        "delta": 5.546
-      }
-    ]
-  },
-  {
-    "from": 6.256,
-    "to": 7.2738,
-    "integral": 7.831
-  }
-]
