--- packages/spectra-data/src/range/visualizer/__tests__/test.js
@@ -1,19 +1,0 @@
-/**
- * Created by acastillo on 3/11/15.
- */
-'use strict';
-
-const Ranges = require('../../Ranges');
-
-const ranges = new Ranges([{from: 7.259458288810639, to: 7.3166664267781405, integral: 1.9236104600745207, signal: [{nbAtoms: 0, diaID: [], multiplicity: 'm', peak: [{x: 7.2641991121231895, intensity: 15889919.418181818, width: 0.0015289194989119607}, {x: 7.2672759891740215, intensity: 52495693.30909091, width: 0.0015614636637360207}, {x: 7.2691750524215575, intensity: 86880073.00441803, width: 0.0032389212036395194}, {x: 7.273581166158989, intensity: 31272505.492640693, width: 0.0025587512701711387}, {x: 7.284596771604648, intensity: 47985190.27532468, width: 0.0028628150018445513}, {x: 7.286982689545268, intensity: 78438220.76883118, width: 0.0026679813948171033}, {x: 7.289080925298546, intensity: 75856273.395671, width: 0.0029787944596295023}, {x: 7.291380273337489, intensity: 36386629.29523809, width: 0.0010423250586938184}, {x: 7.293789374030411, intensity: 17367108.84848485, width: 0.002129074693747238}, {x: 7.301499447724138, intensity: 17204893.056277055, width: 0.002484574471422783}, {x: 7.305619283445369, intensity: 41994063.1981531, width: 0.0029346756599272564}, {x: 7.308006959134621, intensity: 27733650.32727273, width: 0.0028864892145064294}, {x: 7.309918699943227, intensity: 14875206.97835498, width: 0.0007385455382291705}], kind: '', remark: '', delta: 7.282539417551958}], signalID: '1', _highlight: ['1']}, {from: 7.147953157865936, to: 7.234259379921361, integral: 2.8403291428519815, signal: [{nbAtoms: 0, diaID: [], multiplicity: 'm', peak: [{x: 7.155770394578837, intensity: 7775745.440692641, width: 0.002075070956822634}, {x: 7.159481985810329, intensity: 17889350.914108288, width: 0.0038429426481310694}, {x: 7.163095161001141, intensity: 10640563.321212124, width: 0.002193560689379141}, {x: 7.171918922234248, intensity: 10879576.425974026, width: 0.0009232189275686718}, {x: 7.1777024953097985, intensity: 46006873.41355898, width: 0.005942656450177136}, {x: 7.182499070459351, intensity: 18974604.325541127, width: 0.0018868761893605901}, {x: 7.187335699583821, intensity: 10011536.948917748, width: 0.0008130863917514376}, {x: 7.190763286393803, intensity: 28823981.191341992, width: 0.0009896890260689204}, {x: 7.19239031655489, intensity: 48575995.23116884, width: 0.0018887115494769955}, {x: 7.194211924830934, intensity: 65088425.21904763, width: 0.0014733769187005945}, {x: 7.19602145856346, intensity: 74114686.79134199, width: 0.002215725426164812}, {x: 7.198894055941731, intensity: 66609641.579620205, width: 0.004249592509125862}, {x: 7.205903262714266, intensity: 16516264.962770564, width: 0.00036928757102746873}, {x: 7.2073521842814445, intensity: 23365075.72640692, width: 0.000553931356541203}, {x: 7.213567887315607, intensity: 108136286.81714119, width: 0.006897164201918027}, {x: 7.21707015501518, intensity: 71595095.42164503, width: 0.0013138788877665503}, {x: 7.218716342668988, intensity: 55086240.69956712, width: 0.0012370678493707468}], kind: '', remark: '', delta: 7.203205792670877}], signalID: '2', _highlight: ['2']}, {from: 2.5626250911754362, to: 2.6449298824013394, integral: 2.0468838241597087, signal: [{nbAtoms: 0, diaID: [], multiplicity: 'q', peak: [{x: 2.575049332256407, intensity: 40949123.46714861, width: 0.004195204944606687}, {x: 2.5942982570203332, intensity: 113096720.39296657, width: 0.004746599231301991}, {x: 2.613359583638824, intensity: 114783393.40883784, width: 0.004879183709993429}, {x: 2.63218289381467, intensity: 40784291.43920168, width: 0.004695094376783058}], kind: '', remark: '', j: [{coupling: 7.616016909083783, multiplicity: 'q'}], delta: 2.603777486788388}], signalID: '3', _highlight: ['3']}, {from: 1.164466239195155, to: 1.2167547372808154, integral: 2.7967937834841843, signal: [{nbAtoms: 0, diaID: [], multiplicity: 't', peak: [{x: 1.1714104390180176, intensity: 192342923.51719177, width: 0.002356876374308934}, {x: 1.1907370995420268, intensity: 332854981.7724614, width: 0.0029587479981104964}, {x: 1.2095576788578244, intensity: 218114261.40196657, width: 0.002553422826628506}], kind: '', remark: '', j: [{coupling: 7.631020981942299, multiplicity: 't'}], delta: 1.1906104882379853}], signalID: '4', _highlight: ['4']}]);
-
-describe('Annotations 1D from ranges', function () {
-    it('Default options', function () {
-        var annotations = ranges.getAnnotations();
-        annotations.length.should.eql(ranges.length);
-        annotations.forEach(annotation => {
-            annotation._highlight.length.should.eql(1);
-        });
-    });
-});
-
